fix(agents): apply sandbox spec and align CLI args in ClaudeAdapter
ClaudeAdapter::start_turn diverged from ClaudeCodeAgent (claude.rs):

- Never built a SandboxSpec, so capability-token allowed_write_paths
  were silently dropped on the adapter path; now routed through
  spawn_contract::prepare_agent_spawn like claude.rs and codex_adapter.
- Ignored req.env_vars entirely; now propagated via the spawn contract.
- Ignored req.sandbox_mode; now honored (DangerFullAccess default,
  matching codex_adapter).
- Passed --dangerously-skip-permissions together with --allowedTools,
  which are mutually exclusive in Claude CLI 2.1.70+; now uses the
  issue-#514 split (--permission-mode bypassPermissions + --allowedTools
  for restricted profiles).

Test module extracted to claude_adapter_tests.rs (file exceeded the
800-line ceiling) with three new tests: full-profile flag, restricted
profile mutual exclusion, env-var propagation.
